V 2.0

+++ POOB-TEST-FIRST-MASTER-TRAINING-1.1.0/src/main/java/entities/Session.java
@@ -8,7 +8,8 @@ public class Session {
         this.activity = activity;
     }
 
-    public int averageBeats() {
+    public double averageBeats() {
+
         if (activity!= null) {
             return activity.averageBeats();
         }
